% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_clustering.R
\name{cell_clustering}
\alias{cell_clustering}
\title{Cell clustering analysis}
\usage{
cell_clustering(count_data, percent = 0.1, ncores = 4,
  perplexity = 30, need.imputation = FALSE, imputed.data = TRUE)
}
\arguments{
\item{count_data}{The count expression matrix. The rows correspond to genes and
the columns correspond to cells. Can be sparse.}

\item{percent}{Genes that are expressed in less than 100*percent\% of the cells are filtered out. Default is 0.1.}

\item{ncores}{Number of cores to use. Default is 4.}

\item{perplexity}{Perplexity parameter of function Rtsne.}

\item{need.imputation}{Whether the input matrix needs imputation. Default is FALSE.}

\item{imputed.data}{Whether the input matrix has been imputed. Default is TRUE.}
}
\value{
The t-SNE visualization and aRI value of the dataset.
}
\description{
Cell clustering analysis
}
\examples{
data('ipsc_saver')
cell_clustering(ipsc_saver)
}
\author{
Yiqiu Tan
}
