% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deng_saver}
\alias{deng_saver}
\title{Test dataset}
\format{An object of class \code{matrix} with 13514 rows and 286 columns.}
\usage{
deng_saver
}
\description{
A numeric matrix with 13514 rows and 286 columns. This matrix can be
used to test the function \code{trajectory_analysis} provided by the package.
}
\keyword{datasets}
