% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipsc_saver}
\alias{ipsc_saver}
\title{Test dataset}
\format{An object of class \code{matrix} with 11960 rows and 315 columns.}
\usage{
ipsc_saver
}
\description{
A numeric matrix with 11960 rows and 315 columns. This matrix can be
used to test the function \code{cell_clustering} provided by the package.
}
\keyword{datasets}
