% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{log_normalization}
\alias{log_normalization}
\title{Normalization}
\usage{
log_normalization(x)
}
\arguments{
\item{x}{The expression matrix that needs log-normalization.}
}
\value{
The log-normalized expression matrix.
}
\description{
Performs log-normalization on the raw count dataset.
}
\examples{
x = c(1, 2, 3, 4, 5, 6, 7, 8)
\dontrun{
log_normalization(x)
}
}
