% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{preprocessing}
\alias{preprocessing}
\title{Normalization}
\usage{
preprocessing(x, percent)
}
\arguments{
\item{x}{The expression matrix that needs preprocessing or log-normalization.}

\item{percent}{Genes that are expressed in less than 100*percent\% of the cells are filtered out.}
}
\value{
The preprocessed and log-normalized expression matrix.
}
\description{
Performs preprocessing on the raw count dataset.
}
\examples{
x = c(1, 2, 3, 4, 5, 6, 7, 8)
\dontrun{
preprocessing(x)
}
}
