% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_analysis.R
\name{trajectory_analysis}
\alias{trajectory_analysis}
\title{Pseudotime trajectory analysis}
\usage{
trajectory_analysis(count_data, cellLabels, percent = 0.1, ncores = 4,
  need.imputation = FALSE, imputed.data = TRUE)
}
\arguments{
\item{count_data}{The count expression matrix. The rows correspond to genes and
the columns correspond to cells. Can be sparse.}

\item{cellLabels}{The cell labels of the dataset, which is the gold standard of the trajectory analysis.}

\item{percent}{Genes that are expressed in less than 100*percent\% of the cells are filtered out. Default is 0.1.}

\item{ncores}{Number of cores to use. Default is 4.}

\item{need.imputation}{Whether the input matrix needs imputation. Default is FALSE.}

\item{imputed.data}{Whether the input matrix has been imputed. Default is TRUE.}
}
\value{
Two metrics: Pseudo-temporal Ordering Score (POS) and Kendall's rank correlation score,
which are used to assess the accuracy of the inferred cell trajectory and the plot of the cell trajectory.
}
\description{
Carries out cell trajectory analysis to assess the imputation accuracy of SAVER.
}
\examples{
data('deng_saver')
deng_cellLabels <- factor(colnames(deng_saver),
                          levels=c('zygote', 'early 2-cell',
                                   'mid 2-cell', 'late 2-cell',
                                   '4-cell', '8-cell', '16-cell', 'early blastocyst',
                                   'mid blastocyst', 'late blastocyst'))
trajectory_analysis(deng_saver, deng_cellLabels)
}
\author{
Yiqiu Tang
}
